refactor: simplify codebase
