+++ src/lib.rs
@@ -1,6 +1,9 @@
+mod ast;
 mod lexer;
 mod parser;
 mod repl;
+mod utils;
 
-pub use lexer::{Lexer, Token, TokenKind, TokenSpan, TokenSpanTracker};
+pub use ast::*;
+pub use lexer::Lexer;
 pub use repl::REPL;
